+++ packages/medusa/src/api/auth/utils/generate-jwt-token.ts
@@ -1,5 +1,7 @@
 import {
   AuthIdentityDTO,
+  AuthMfaChallengeDTO,
+  ConfigModule,
   MedusaContainer,
   ProjectConfigOptions,
 } from "@medusajs/framework/types"
@@ -10,6 +12,86 @@ import {
 } from "@medusajs/framework/utils"
 import { type Secret } from "jsonwebtoken"
 import RbacFeatureFlag from "../../../feature-flags/rbac"
+import { validateVerification } from "./validate-verification"
+
+export async function generateJwtTokenWithChecks(
+  container: MedusaContainer,
+  {
+    authIdentity,
+    mfaChallenge,
+    actorType,
+    authProvider,
+  }: {
+    authIdentity: AuthIdentityDTO
+    mfaChallenge?: AuthMfaChallengeDTO
+    actorType: string
+    authProvider: string
+  }
+) {
+  const config: ConfigModule = container.resolve(
+    ContainerRegistrationKeys.CONFIG_MODULE
+  )
+  const { http } = config.projectConfig
+
+  const actorlessToken = await generateJwtTokenForAuthIdentity(
+    {
+      authIdentity,
+      actorType: actorType,
+      authProvider: authProvider,
+      container: container,
+    },
+    {
+      secret: http.jwtSecret!,
+      expiresIn: http.jwtExpiresIn,
+      // Running a verification is about the auth identity, so we return a token to be able to authenticate the requests
+      // without having an actor tied to it until the verification is completed.
+      skipActorType: true,
+      options: http.jwtOptions,
+    }
+  )
+
+  // Check if verification of the provider entity data is required (such as email verification)
+  const { requiresVerification, verification } = await validateVerification(
+    container,
+    {
+      actor_type: actorType,
+      auth_provider: authProvider,
+      auth_identity: authIdentity,
+    }
+  )
+
+  if (requiresVerification && !verification?.verified_at) {
+    return {
+      verification_required: true,
+      verification,
+      token: actorlessToken,
+    }
+  }
+
+  if (mfaChallenge) {
+    return {
+      mfa_required: true,
+      mfa_challenge: mfaChallenge,
+      token: actorlessToken,
+    }
+  }
+
+  const token = await generateJwtTokenForAuthIdentity(
+    {
+      authIdentity,
+      actorType: actorType,
+      authProvider: authProvider,
+      container,
+    },
+    {
+      secret: http.jwtSecret!,
+      expiresIn: http.jwtExpiresIn,
+      options: http.jwtOptions,
+    }
+  )
+
+  return { token }
+}
 
 export async function generateJwtTokenForAuthIdentity(
   {
@@ -26,18 +108,20 @@ export async function generateJwtTokenForAuthIdentity(
   {
     secret,
     expiresIn,
+    skipActorType,
     options,
   }: {
     secret: Secret
     expiresIn: string | undefined
+    skipActorType?: boolean
     options?: ProjectConfigOptions["http"]["jwtOptions"]
   }
 ) {
   const expiresIn_ = expiresIn ?? options?.expiresIn
   const entityIdKey = `${actorType}_id`
-  const entityId = authIdentity?.app_metadata?.[entityIdKey] as
-    | string
-    | undefined
+  const entityId = skipActorType
+    ? undefined
+    : (authIdentity?.app_metadata?.[entityIdKey] as string | undefined)
 
   const providerIdentity = !authProvider
     ? undefined