+++ packages/plugins/loyalty/src/admin/routes/store-credit-accounts/[id]/page.tsx
@@ -1,27 +1,30 @@
-import { useParams } from "react-router-dom";
-import { TwoColumnLayout } from "../../../components/layouts/two-column";
-import { useStoreCreditAccount } from "../../../hooks/api/store-credit-accounts";
-import StoreCreditAccountBalanceSection from "./components/store-credit-account-balance-section";
-import StoreCreditAccountDetailsSection from "./components/store-credit-account-details-section";
-import StoreCreditAccountCustomerSection from "./components/store-credit-account-customer-section";
-import { TransactionsTable } from "./components/transactions-table/table";
-import StoreCreditAccountCodeSection from "./components/store-credit-account-code";
+import { useParams } from "react-router-dom"
+import { LayoutComposer } from "@medusajs/dashboard/components"
+import { useStoreCreditAccount } from "../../../hooks/api/store-credit-accounts"
+import StoreCreditAccountBalanceSection from "./components/store-credit-account-balance-section"
+import StoreCreditAccountDetailsSection from "./components/store-credit-account-details-section"
+import StoreCreditAccountCustomerSection from "./components/store-credit-account-customer-section"
+import { TransactionsTable } from "./components/transactions-table/table"
+import StoreCreditAccountCodeSection from "./components/store-credit-account-code"
 
 const StoreCreditAccountPage = () => {
-  const { id } = useParams();
+  const { id } = useParams()
 
   const { store_credit_account: storeCreditAccount } = useStoreCreditAccount(
     id!
-  );
+  )
 
   if (!storeCreditAccount) {
-    return;
+    return
   }
 
   return (
-    <>
-      <TwoColumnLayout
-        firstCol={
+    <LayoutComposer
+      widgetsZonePrefix="store_credit_account.details"
+      preferredLayoutId="core:two-column"
+      data={storeCreditAccount}
+      sections={{
+        main: (
           <>
             <StoreCreditAccountDetailsSection
               storeCreditAccount={storeCreditAccount}
@@ -29,8 +32,8 @@ const StoreCreditAccountPage = () => {
 
             <TransactionsTable id={storeCreditAccount.id} />
           </>
-        }
-        secondCol={
+        ),
+        side: (
           <>
             <StoreCreditAccountBalanceSection
               storeCreditAccount={storeCreditAccount}
@@ -40,10 +43,10 @@ const StoreCreditAccountPage = () => {
               customerId={storeCreditAccount.customer_id}
             />
           </>
-        }
-      />
-    </>
-  );
-};
+        ),
+      }}
+    />
+  )
+}
 
-export default StoreCreditAccountPage;
+export default StoreCreditAccountPage